Crawler_Tax_Portal/requirements.txt
# GUI Components (PyQt5)
PyQt5==5.15.9
PyQt5-Qt5
PyQt5-sip
PyQt5Designer

# Web Framework
Flask==3.0.0
flask-cors==5.0.0

# Database
pymongo==4.12.0

# HTTP Requests & Web Scraping
requests==2.31.0
urllib3==2.0.7
beautifulsoup4
selenium
selenium-print

# Data Processing
pandas==2.2.3
openpyxl==3.1.2
python-dateutil==2.8.2

# System & Utilities
python-dotenv==1.0.0
pyhtml2pdf==0.0.1

# Windows Specific (có thể bỏ qua cho Docker)
pywin32
wmi==1.5.1
